Upgrade cabal to 3.10 branch
Switch to 3.10 branch for cabal submodule

Update internal libraries in hackport.cabal to match their upstream
counterparts. Module names were re-ordered to match their ordering
upstream.

NOTE: this introduces a new lower bound for a dependency:
`directory >=1.3.7.0`. This is not bundled with GHC until v9.4, which
will cause problems for Gentoo users unless a) they upgrade to newer GHC
or b) the 'directory' library is upgraded for older GHC versions in the
tree.

Signed-off-by: hololeap <hololeap@protonmail.com>